requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
cffi==1.12.3
cryptography==2.7
Django==2.2.5
djangorestframework==3.10.3
Fabric3==1.14.post1
paramiko==2.6.0
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
uWSGI==2.0.18
